Renamed service
- Renamed utils.py into service.py
