+++ app/services/db/common.py
@@ -4,7 +4,7 @@ Session module
 import mysql
 from mysql.connector import Error
 
-from app.utils.logger.actions import LoggerService
+from app.services.logger.actions import LoggerService
 
 logger = LoggerService().logger
 class DbConnection:
