+++ app/services/user/actions.py
@@ -11,9 +11,9 @@ from sqlalchemy import or_
 from app.core.access_token import generate_user_token_and_return_user, decode_access_token
 from app.db.models.users import User
 from app.email.email_service import EmailService, EmailSchema
-from app.utils.audit.actions import log_audit_event
-from app.utils.error.user import UserErrorHandler
-from app.utils.logger.actions import LoggerService
+from app.services.audit.actions import log_audit_event
+from app.services.error.user import UserErrorHandler
+from app.services.logger.actions import LoggerService
 
 logger = LoggerService().logger
